Refactored path tracing to sample multiple paths per pixel and allow for continuous rendering over multiple frames. Cornell Box looks a bit bland now, did I brake something?

+++ AwesomeRenderer/random.h
@@ -18,7 +18,7 @@ namespace AwesomeRenderer
 
 		}
 
-		float NextFloat() { return rand() / (float)RAND_MAX; }
+		float NextFloat() { return ((uint32_t)rand()) / (float)RAND_MAX; }
 	};
 
 }
